[Silver II] Title: 예산, Time: 4 ms, Memory: 2152 KB -BaekjoonHub

+++ 백준/Silver/2512. 예산/예산.cc
@@ -6,8 +6,8 @@ using namespace std;
 
 int N;
 vector<int> v;
-long long sum;
-long long cur_sum;
+int sum;
+int cur_sum;
 
 
 void input(){
